wp version and plugin updated

+++ blog/wp-content/plugins/wp-user-avatar/src/Membership/PaymentMethods/Stripe/WebhookHelpers.php
@@ -3,6 +3,8 @@
 namespace ProfilePress\Core\Membership\PaymentMethods\Stripe;
 
 use ProfilePress\Core\Membership\PaymentMethods\Stripe\WebhookHandlers\ChargeRefunded;
+use ProfilePress\Core\Membership\PaymentMethods\Stripe\WebhookHandlers\CheckoutSessionAsyncPaymentFailed;
+use ProfilePress\Core\Membership\PaymentMethods\Stripe\WebhookHandlers\CheckoutSessionAsyncPaymentSucceeded;
 use ProfilePress\Core\Membership\PaymentMethods\Stripe\WebhookHandlers\CheckoutSessionCompleted;
 use ProfilePress\Core\Membership\PaymentMethods\Stripe\WebhookHandlers\CustomerSubscriptionCreated;
 use ProfilePress\Core\Membership\PaymentMethods\Stripe\WebhookHandlers\CustomerSubscriptionDeleted;
@@ -15,13 +17,15 @@ class WebhookHelpers
     public static function valid_events()
     {
         return apply_filters('ppress_stripe_webhooks_whitelist', [
-            'checkout.session.completed'    => new CheckoutSessionCompleted(),
-            'customer.subscription.created' => new CustomerSubscriptionCreated(),
-            'customer.subscription.updated' => new CustomerSubscriptionUpdated(),
-            'customer.subscription.deleted' => new CustomerSubscriptionDeleted(), //triggers when sub is cancelled
-            'invoice.payment_succeeded'     => new InvoicePaymentSucceeded(),
-            'payment_intent.succeeded'      => new PaymentIntentSucceeded(),
-            'charge.refunded'               => new ChargeRefunded()
+            'checkout.session.completed'               => new CheckoutSessionCompleted(),
+            'checkout.session.async_payment_succeeded' => new CheckoutSessionAsyncPaymentSucceeded(),
+            'checkout.session.async_payment_failed'    => new CheckoutSessionAsyncPaymentFailed(),
+            'customer.subscription.created'            => new CustomerSubscriptionCreated(),
+            'customer.subscription.updated'            => new CustomerSubscriptionUpdated(),
+            'customer.subscription.deleted'            => new CustomerSubscriptionDeleted(), //triggers when sub is cancelled
+            'invoice.payment_succeeded'                => new InvoicePaymentSucceeded(),
+            'payment_intent.succeeded'                 => new PaymentIntentSucceeded(),
+            'charge.refunded'                          => new ChargeRefunded()
         ]);
     }
 
@@ -147,14 +151,21 @@ class WebhookHelpers
     /**
      * Creates a Stripe webhook endpoint with the current plugin and site settings.
      *
-     * @return array|false|mixed
+     * @param $endpoint_id
+     * @param bool $without_persist
+     *
+     * @return void
      * @throws \Exception
      */
-    public static function delete($endpoint_id)
+    public static function delete($endpoint_id, $without_persist = false)
     {
-        $endpoint = APIClass::stripeClient()->webhookEndpoints->delete($endpoint_id);
+        if ( ! empty($endpoint_id)) {
+            $endpoint = APIClass::stripeClient()->webhookEndpoints->delete($endpoint_id);
 
-        self::persist($endpoint, true);
+            if ( ! $without_persist) {
+                self::persist($endpoint, true);
+            }
+        }
     }
 
     /**
